[EFL] Fix inappropriate move semantics use in parseAndRemoveEnvironments function https://bugs.webkit.org/show_bug.cgi?id=150544
Patch by Joonghun Park <jh718.park@samsung.com> on 2015-11-01
Reviewed by Darin Adler.

In parseAndRemoveEnvironments function, it declares its paramerter type as rvalue.
But in ProcessLauncher::launchProcess(), the argument of
parseAndRemoveEnvironments still be used after the call of the function.
Currently the rvalue is not assigned to any other variable in parseAndRemoveEnvironments,
so it doesn't lead to runtime error but use of move semantic
is not appropriate semantically in this case.

* UIProcess/Launcher/efl/ProcessLauncherEfl.cpp:
(WebKit::parseAndRemoveEnvironments):
(WebKit::ProcessLauncher::launchProcess):

Identifier: 168960@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191868 268f45cc-cd09-0410-ab3c-d52691b4dbfc