[Win] MiniBrowser unable to use WebInspector https://bugs.webkit.org/show_bug.cgi?id=150810 <rdar://problem/23358514>
Reviewed by Timothy Hatcher.

The CMakeList rule for creating the InjectedScriptSource.min.js was improperly including
the quote characters in the text prepended to InjectedScriptSource.min.js. This caused a
parsing error in the JS file.

The solution was to switch from using "COMMAND echo" to use the more cross-platform
compatible command "COMMAND ${CMAKE_COMMAND} -E echo ...", which handles the string
escaping properly on all platforms.

* CMakeLists.txt: Switch the 'echo' command syntax to be more cross-platform.


Identifier: 169007@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191917 268f45cc-cd09-0410-ab3c-d52691b4dbfc